Try to improve the makefiles a bit
Add some targets to the makefile in the src/ directory:
 - "make ARCH=I386 OS=Linux all" does now something (lazily)
 - "make boostrap" forces a full boostrap (non-lazy)
 - "make help" is now the default target explaining the most imporant things

* src/Makefile: Add the above mentioned targets

* src/images/GNUMake ($(RUNTIME)): Make RUNTIME itself a target
and make various of the .img rules dependent on it.

(compile-pervasive, build-basis, build-batch, build-interpreter,
build-xinterpreter): New phony targets.  Those invoke the batch
compiler on the .mlp files which does proper comparision of file
modification times.  These targets are then used as prerequisites of
the .img files.

(boot-batch.img): New target. This creates the batch compiler that is
used to compile the other things.